requirements.txt
aiofiles==0.6.0
aiosqlite==0.17.0
pytz==2021.1
PyYAML==5.4.1
python-dotenv==0.19.0
pyfiglet==0.8.post1
asyncpraw==7.4.0
python-dateutil==2.8.2
arrow==1.2.1
disnake==2.2.1
rapidfuzz==1.8.3
humanize==3.12.0
Pillow==8.3.2